requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
greenlet==3.1.1
h11==0.14.0
identify==2.6.9
idna==3.10
nodeenv==1.9.1
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
virtualenv==20.30.0
